iss #22: multi project build, upgrade deps, add methodParamIndentSize to scalastyle

+++ mt8057-agent/src/main/scala/ru/maizy/ambient7/mt8057agent/DeviceEmulator.scala
@@ -41,7 +41,7 @@ class DeviceEmulator(private val queue: JavaQueue[Event]) {
             queue.add(TempUpdated(Temp(lastTemp), currentNanoTime()))
           }
         }
-        Thread.sleep(500 + randomGen.nextInt(500))
+        Thread.sleep((500 + randomGen.nextInt(500)).toLong)
       }
     }
   }
